add update and delete methods

+++ app/repository/base_crud.py
@@ -21,8 +21,8 @@ class CRUDBase(Generic[ModelType, CreateSchemaType, UpdateSchemaType]):
         """
         self.model = model
 
-    def get(self, db: Session, id: Any) -> Optional[ModelType]:
-        return db.query(self.model).filter(self.model.id == id).first()
+    def get(self, *, db: Session, uuid: Any) -> Optional[ModelType]:
+        return db.query(self.model).filter(self.model.uuid == uuid).first()
 
     def get_multi(
         self, db: Session, *, skip: int = 0, limit: int = 100
@@ -57,8 +57,12 @@ class CRUDBase(Generic[ModelType, CreateSchemaType, UpdateSchemaType]):
         db.refresh(db_obj)
         return db_obj
 
-    def remove(self, db: Session, *, id: int) -> ModelType:
-        obj = db.query(self.model).get(id)
+    def remove(self, *, db: Session, uuid: Any) -> ModelType:
+        obj = db.query(self.model).filter(self.model.uuid == uuid).first()
+
+        if not obj:
+            return None
+        
         db.delete(obj)
         db.commit()
         return obj
